backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
PyMySQL==1.1.0
cryptography==41.0.7
Werkzeug==3.0.1
marshmallow==3.20.1
python-dotenv==1.0.0
gunicorn==21.2.0